requirements.txt
pandas>=1.5
numpy>=1.23
statsmodels>=0.14
matplotlib>=3.6
seaborn>=0.12
streamlit>=1.12
plotly>=5.24
streamlit-extras>=0.4.0
streamlit-option-menu>=0.3.12
streamlit-aggrid>=0.3.4.post3
prophet>=1.1
lightgbm>=4.0
tbats>=1.1
neuralprophet>=0.6
python-pptx>=0.6.21
python-docx>=1.1.2
xlsxwriter>=3.2.0
reportlab>=4.0.4
fpdf2>=2.7.4
svglib>=1.5.1
pytest>=8.2
